Move select timeout into main SD code, Add timeout to test_sd
