+++ tests/requirements.txt
@@ -19,6 +19,7 @@
 # same modules should appear in the setup.py list as given below.
 pytest==8.3.3
 pytest-xdist==3.6.1
+pytest-timeout==2.4.0
 filelock==3.19.1
 numpy==2.3.2
 scipy==1.16.1
