+++ tests/test_pwm_idle/run_sim.sh
