+++ tests/test_pwm_idle/sim.gtkw
@@ -1,15 +1,15 @@
 [*]
 [*] GTKWave Analyzer v3.3.116 (w)1999-2023 BSI
-[*] Wed Sep 17 16:53:25 2025
+[*] Wed Sep 17 16:56:41 2025
 [*]
 [dumpfile] "/home/ed/sandboxes/sw_dac/lib_sw_dac/tests/test_stuff/trace.vcd"
-[dumpfile_mtime] "Wed Sep 17 16:51:52 2025"
+[dumpfile_mtime] "Wed Sep 17 16:54:52 2025"
 [dumpfile_size] 16117009
 [savefile] "/home/ed/sandboxes/sw_dac/lib_sw_dac/tests/test_stuff/sim.gtkw"
-[timestart] 80990000
+[timestart] 64250000
 [size] 1580 1162
-[pos] -1 -1
-*-23.490168 109320000 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1
+[pos] -292 -36
+*-24.090168 109320000 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1
 [sst_width] 463
 [signals_width] 342
 [sst_expanded] 1
@@ -19,13 +19,9 @@ tile[0]_XS1_PORT_1A
 tile[0]_XS1_PORT_1B
 @22
 tile[0]_ClkBlk1_fallEdge
-tile[0]_Core0_waitingResourceId[31:0]
-@820
-tile[0]_Core0_waitingResourceName[63:0]
-@22
 tile[0]_Core1_waiting
 tile[0]_Core1_waitingResourceId[31:0]
-@820
+@821
 tile[0]_Core1_waitingResourceName[63:0]
 @22
 tile[0]_INSTRUCTION_PC[31:0]
@@ -34,10 +30,14 @@ tile[0]_INSTRUCTION_THREAD_ID[7:0]
 tile[0]_INSTRUCTION_MNEMONIC[63:0]
 @22
 tile[0]_Core0_waiting
+@820
+tile[0]_Core0_waitingResourceName[63:0]
+@22
+tile[0]_Core0_waitingResourceId[31:0]
 tile[0]_Timer0_time[63:0]
 @24
 tile[0]_CYCLES[31:0]
-@23
+@22
 tile[0]_XS1_PORT_1A_transferReg[31:0]
 [pattern_trace] 1
 [pattern_trace] 0
