+++ tests/test_pwm_idle/src/config.xscope
